# services/media-service/docs/swagger.yaml
basePath: /media
definitions:
  httpapi.assetResponse:
    properties:
      bucket:
        type: string
      contentType:
        type: string
      createdAt:
        type: string
      durationSeconds:
        type: integer
      height:
        type: integer
      id:
        type: string
      mediaType:
        type: string
      objectKey:
        type: string
      publicUrl:
        type: string
      sizeBytes:
        type: integer
      status:
        type: string
      uploadedBy:
        type: string
      width:
        type: integer
    type: object
  httpapi.attachmentResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      mediaAssetId:
        type: string
      ownerId:
        type: string
      ownerService:
        type: string
      ownerType:
        type: string
      position:
        type: integer
      role:
        type: string
      updatedAt:
        type: string
    type: object
  httpapi.bulkAssetItem:
    properties:
      asset:
        $ref: '#/definitions/httpapi.assetResponse'
      found:
        type: boolean
      id:
        type: string
    type: object
  httpapi.bulkAssetsByOwnerRoleItem:
    properties:
      asset:
        $ref: '#/definitions/httpapi.assetResponse'
      found:
        type: boolean
      ownerId:
        type: string
      role:
        type: string
    type: object
  httpapi.bulkAssetsByOwnerRoleRequest:
    properties:
      items:
        items:
          $ref: '#/definitions/httpapi.ownerRolePairRequest'
        type: array
      ownerService:
        type: string
      ownerType:
        type: string
    type: object
  httpapi.bulkAssetsByOwnerRoleResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/httpapi.bulkAssetsByOwnerRoleItem'
        type: array
    type: object
  httpapi.bulkAssetsRequest:
    properties:
      assetIds:
        items:
          type: string
        type: array
    type: object
  httpapi.bulkAssetsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/httpapi.bulkAssetItem'
        type: array
    type: object
  httpapi.confirmUploadRequest:
    properties:
      durationSeconds:
        type: integer
      height:
        type: integer
      width:
        type: integer
    type: object
  httpapi.createAttachmentRequest:
    properties:
      mediaAssetId:
        type: string
      ownerId:
        type: string
      ownerService:
        type: string
      ownerType:
        type: string
      position:
        type: integer
      role:
        type: string
    type: object
  httpapi.errorResponse:
    properties:
      error:
        type: string
    type: object
  httpapi.ownerRolePairRequest:
    properties:
      ownerId:
        type: string
      role:
        type: string
    type: object
  httpapi.requestUploadRequest:
    properties:
      checksum:
        type: string
      contentType:
        type: string
      mediaType:
        type: string
    type: object
  httpapi.requestUploadResponse:
    properties:
      assetId:
        type: string
      bucket:
        type: string
      expiresInSeconds:
        type: integer
      objectKey:
        type: string
      uploadUrl:
        type: string
    type: object
info:
  contact: {}
  title: Media Service API
  version: "1.0"
paths:
  /assets/{id}:
    get:
      description: Get detail information of media asset by ID
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.assetResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Get asset
      tags:
      - Asset
  /assets/bulk:
    post:
      consumes:
      - application/json
      description: Get detail information for multiple media assets in one call. Response
        items preserve the exact order of the requested assetIds, including duplicates
        and not-found ids (found=false)
      parameters:
      - description: List of asset IDs (max 100)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/httpapi.bulkAssetsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.bulkAssetsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Get multiple assets
      tags:
      - Asset
  /assets/by-owner-role/bulk:
    post:
      consumes:
      - application/json
      description: Resolve multiple (ownerId, role) pairs to their currently-attached
        asset in one call — e.g. fetch the thumbnail for many products on a listing
        page. Response items preserve the exact order of the requested items, including
        pairs that resolve to nothing (found=false)
      parameters:
      - description: Owner service/type + list of (ownerId, role) pairs (max 100)
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/httpapi.bulkAssetsByOwnerRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.bulkAssetsByOwnerRoleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Get assets by owner+role
      tags:
      - Asset
  /attachments:
    get:
      description: Get attachment list, filter by owner_service, owner_type, owner_id,
        role
      parameters:
      - description: Owner service
        in: query
        name: owner_service
        type: string
      - description: Owner type
        in: query
        name: owner_type
        type: string
      - description: Owner ID
        in: query
        name: owner_id
        type: string
      - description: Role
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.attachmentResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: List attachments
      tags:
      - Attachment
    post:
      consumes:
      - application/json
      description: Assign media asset to specific entity (owner)
      parameters:
      - description: Create attachment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.createAttachmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.attachmentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Create attachment
      tags:
      - Attachment
  /attachments/{id}:
    delete:
      description: Delete attachment by ID
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Delete attachment
      tags:
      - Attachment
  /uploads:
    post:
      consumes:
      - application/json
      description: Create presigned URL to upload media
      parameters:
      - description: Upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.requestUploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.requestUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Request upload
      tags:
      - Upload
  /uploads/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Confirm upload completed and update metadata (size, duration...)
        for asset
      parameters:
      - description: Asset ID
        in: path
        name: id
        required: true
        type: string
      - description: Confirm upload request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httpapi.confirmUploadRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.assetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Confirm upload
      tags:
      - Upload
swagger: "2.0"